learn/05-orchestration/pyproject.toml
[project]
name = "05-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.4",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
]
